+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 3.15.3 - 2024-11-18
+
 - fix: mangled wireframe layouts ([#250](https://github.com/PostHog/posthog-ios/pull/250))
 - recording: do not rotate the session id for hybrid SDKs ([#253](https://github.com/PostHog/posthog-ios/pull/253))